requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
vnstock==3.3.1
pandas==2.2.0
numpy==1.26.3
ta==0.11.0
redis==5.0.1
google-generativeai==0.8.0
python-multipart==0.0.9
pydantic==2.5.3
aiohttp==3.9.1
python-dotenv==1.0.0
firebase-admin==6.4.0
